refactor today question component
