+++ app/components/todayquestion-components/todayquestion.tsx
@@ -2,111 +2,8 @@
 
 import { useEffect, useState } from 'react';
 import { CheckCircleIcon, XCircleIcon, ExternalLinkIcon, CalendarIcon } from '@/app/components/todayquestion-components/icons';
-
-interface TodayQuestion {
-    first_question_id: string | null;
-    first_question_url: string | null;
-    first_question_solved: string | null;
-    second_question_id: string | null;
-    second_question_url: string | null;
-    second_question_solved: string | null;
-}
-
-interface ApiResponse {
-    status: string;
-    message?: string;
-    first_question_id: string | null;
-    first_question_url: string | null;
-    first_question_solved: string | null;
-    second_question_id: string | null;
-    second_question_url: string | null;
-    second_question_solved: string | null;
-}
-
-interface QuestionCardProps {
-    questionNumber: string;
-    questionId: string | null;
-    questionUrl: string | null;
-    questionSolved: string | null;
-}
-
-const QuestionCard = ({ questionNumber, questionId, questionUrl, questionSolved }: QuestionCardProps) => {
-    const isSolved = questionSolved === 'true';
-    const hasData = questionId && questionUrl;
-
-    const getTitle = (url: string | null) => {
-        if (!url) return 'No Question Available';
-        const title = url.split('/problems/')[1]?.split('/')[0]?.replace(/-/g, ' ');
-        return title ? title.charAt(0).toUpperCase() + title.slice(1) : 'Problem';
-    };
-
-    if (!hasData) {
-        return (
-            <div className="bg-white rounded-lg shadow-md p-6 border-l-4 border-gray-400">
-                <div className="flex items-center justify-between mb-4">
-                    <h3 className="text-lg font-semibold text-gray-800">{questionNumber} Question</h3>
-                    <div className="flex items-center text-gray-400">
-                        <XCircleIcon className="w-5 h-5 mr-2" />
-                        <span className="text-sm">Not Available</span>
-                    </div>
-                </div>
-                <div className="text-center py-8 text-gray-500">
-                    <CalendarIcon className="w-8 h-8 mx-auto mb-2 opacity-50" />
-                    <p className="text-sm">No question data available for today</p>
-                </div>
-            </div>
-        );
-    }
-
-    return (
-        <div className={`bg-white rounded-lg shadow-md p-6 border-l-4 ${isSolved ? 'border-green-500' : 'border-orange-500'}`}>
-            <div className="flex items-center justify-between mb-4">
-                <h3 className="text-lg font-semibold text-gray-800">{questionNumber} Question</h3>
-                <div className={`flex items-center ${isSolved ? 'text-green-600' : 'text-orange-600'}`}>
-                    {isSolved ? (
-                        <>
-                            <CheckCircleIcon className="w-5 h-5 mr-2" />
-                            <span className="text-sm font-medium">Solved</span>
-                        </>
-                    ) : (
-                        <>
-                            <XCircleIcon className="w-5 h-5 mr-2" />
-                            <span className="text-sm font-medium">Not Solved</span>
-                        </>
-                    )}
-                </div>
-            </div>
-
-            <div className="space-y-3">
-                <div>
-                    <p className="text-sm text-gray-600 mb-1">Problem Title:</p>
-                    <p className="font-medium text-gray-900 capitalize">{getTitle(questionUrl)}</p>
-                </div>
-
-                <div>
-                    <p className="text-sm text-gray-600 mb-1">Question ID:</p>
-                    <p className="text-gray-700">#{questionId}</p>
-                </div>
-
-                <div className="pt-2">
-                    <a
-                        href={questionUrl || '#'}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className={`inline-flex items-center px-4 py-2 rounded-md text-sm font-medium transition-colors ${
-                            isSolved
-                                ? 'bg-green-100 text-green-800 hover:bg-green-200'
-                                : 'bg-orange-100 text-orange-800 hover:bg-orange-200'
-                        }`}
-                    >
-                        <ExternalLinkIcon className="w-4 h-4 mr-2" />
-                        {isSolved ? 'Review Problem' : 'Solve Problem'}
-                    </a>
-                </div>
-            </div>
-        </div>
-    );
-};
+import {TodayQuestion, QuestionCardProps, ApiResponse} from "@/app/components/todayquestion-components/todayquuestion-interface";
+import { QuestionCard } from '@/app/components/todayquestion-components/questioncard';
 
 export default function TodaysQuestions() {
     const [todayData, setTodayData] = useState<TodayQuestion | null>(null);
@@ -117,7 +14,7 @@ export default function TodaysQuestions() {
         async function fetchTodaysQuestions() {
             try {
                 const response = await fetch('/api/fetch-today-question', {
-                    cache: 'no-store' // Since your API sets no-cache headers
+                    cache: 'no-store'
                 });
 
                 if (!response.ok) {
@@ -142,7 +39,6 @@ export default function TodaysQuestions() {
                 console.error('Error fetching today\'s questions:', err);
                 setError(err instanceof Error ? err.message : 'Unknown error occurred');
 
-                // Fallback mock data for development
                 setTodayData({
                     first_question_id: '1',
                     first_question_url: 'https://leetcode.com/problems/two-sum/',
